requirements.txt
pandas==2.0
pyarrow==11.0
numpy==1.24
sentence-transformers==2.2.2
huggingface-hub==0.21.0
tqdm==4.65
sqlalchemy==2.0
pyyaml==6.0
kagglehub==0.3.0
scikit-learn==1.2
accelerate==0.33.0
filelock==3.16.0
requests==2.31.0
flashtext2
uvicorn
fastapi
qdrant-client==1.14.2
langchain==0.3.21
langchain-huggingface==0.1.2
langchain-qdrant==0.2.0
